modularized the suppliers page
