+++ src/constants/modules.ts
@@ -1,0 +1,14 @@
+import { type ModulesProps } from "../lib/types";
+
+export const modules: ModulesProps[] = [
+  {
+    title: "Suppliers",
+    description: "Manage the suppliers for the system.",
+    href: "/suppliers",
+  },
+  {
+    title: "Users",
+    description: "Manage the users in the system.",
+    href: "/users",
+  },
+];
