+++ src/pages/Suppliers/components/SupplierHeader.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Typography } from "@mui/material";
+
+const SuppliersHeader: React.FC = () => {
+  return (
+    <Typography variant="h4" gutterBottom>
+      Suppliers
+    </Typography>
+  );
+};
+
+export default SuppliersHeader;
